Rename module to compaction
